+++ script/bench-c-deflate-fragmentation.c
@@ -61,6 +61,7 @@
 // If building this program in an environment that doesn't easily accomodate
 // relative includes, you can use the script/inline-c-relative-includes.go
 // program to generate a stand-alone C file.
+#include "../gen/c/std/adler32.c"
 #include "../gen/c/std/deflate.c"
 #include "../gen/c/std/zlib.c"